bug fix and code refactor
